Refactor files in the type engine.
