+++ sway-core/src/parse_tree/declaration/type_parameter.rs
@@ -1,8 +1,4 @@
-use crate::{
-    semantic_analysis::{CopyTypes, TypeMapping},
-    type_engine::*,
-    CallPath, TypedDeclaration,
-};
+use crate::{error::*, parse_tree::*, semantic_analysis::*, type_engine::*};
 
 use sway_types::{ident::Ident, span::Span};
 
@@ -59,6 +55,33 @@ impl CopyTypes for TypeParameter {
                 insert_type(ty)
             }
         };
+    }
+}
+
+impl UpdateTypes for TypeParameter {
+    fn update_types(
+        &mut self,
+        type_mapping: &TypeMapping,
+        namespace: &mut Namespace,
+        self_type: TypeId,
+    ) -> CompileResult<()> {
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        self.type_id = match look_up_type_id(self.type_id).matches_type_parameter(type_mapping) {
+            Some(matching_id) => insert_type(TypeInfo::Ref(matching_id, self.span())),
+            None => check!(
+                namespace.resolve_type_with_self(
+                    look_up_type_id(self.type_id),
+                    self_type,
+                    &self.span(),
+                    EnforceTypeArguments::Yes
+                ),
+                insert_type(TypeInfo::ErrorRecovery),
+                warnings,
+                errors,
+            ),
+        };
+        ok((), warnings, errors)
     }
 }
