+++ sway-core/src/semantic_analysis/ast_node/declaration/struct.rs
@@ -1,16 +1,9 @@
 use crate::{
-    error::*,
-    namespace::Items,
-    parse_tree::*,
-    semantic_analysis::{ast_node::copy_types::TypeMapping, insert_type_parameters, CopyTypes},
-    type_engine::*,
-    CompileError, CompileResult, Ident, Namespace,
+    error::*, namespace::*, parse_tree::*, semantic_analysis::*, type_engine::*, types::*,
 };
 use fuels_types::Property;
 use std::hash::{Hash, Hasher};
-use sway_types::Span;
-
-use super::{monomorphize_inner, CreateTypeId, EnforceTypeArguments, MonomorphizeHelper};
+use sway_types::{Ident, Span};
 
 #[derive(Clone, Debug, Eq)]
 pub struct TypedStructDeclaration {
