+++ sway-core/src/semantic_analysis/ast_node/declaration/variable.rs
@@ -1,11 +1,4 @@
-use crate::{
-    constants::*,
-    error::*,
-    semantic_analysis::{CopyTypes, TypeMapping, TypedExpression},
-    type_engine::TypeId,
-    type_engine::*,
-    Ident, Visibility,
-};
+use crate::{constants::*, error::*, semantic_analysis::*, type_engine::*, Ident, Visibility};
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum VariableMutability {
