+++ sway-core/src/semantic_analysis/ast_node/expression/mod.rs
@@ -3,6 +3,6 @@ mod struct_expr_field;
 pub mod typed_expression;
 pub mod typed_expression_variant;
 
-pub(crate) use struct_expr_field::TypedStructExpressionField;
-pub(crate) use typed_expression::{error_recovery_expr, TypedExpression};
+pub(crate) use struct_expr_field::*;
+pub(crate) use typed_expression::*;
 pub(crate) use typed_expression_variant::*;
