+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression.rs
@@ -7,8 +7,6 @@ mod struct_field_access;
 mod tuple_index_access;
 mod unsafe_downcast;
 
-use std::collections::{HashMap, VecDeque};
-
 pub(crate) use self::{
     enum_instantiation::*, function_application::*, if_expression::*, lazy_operator::*,
     method_application::*, struct_field_access::*, tuple_index_access::*, unsafe_downcast::*,
@@ -16,13 +14,11 @@ pub(crate) use self::{
 
 use super::match_expression::{check_match_expression_usefulness, TypedMatchExpression};
 
-use crate::{
-    semantic_analysis::ast_node::*,
-    type_engine::TypeId,
-    type_engine::{insert_type, AbiName, IntegerBits},
-};
+use crate::{error::*, parse_tree::*, semantic_analysis::*, type_engine::*};
 
-use ast_node::declaration::CreateTypeId;
+use sway_types::{Ident, Span};
+
+use std::collections::{HashMap, VecDeque};
 
 #[derive(Clone, Debug, Eq)]
 pub struct TypedExpression {
@@ -1875,6 +1871,7 @@ impl TypedExpression {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::types::*;
 
     fn do_type_check(expr: Expression, type_annotation: TypeId) -> CompileResult<TypedExpression> {
         let mut namespace = Namespace::init_root(namespace::Module::default());
