+++ sway-core/src/semantic_analysis/namespace/root.rs
@@ -1,11 +1,6 @@
 use crate::{
-    error::*,
-    semantic_analysis::{
-        ast_node::TypedExpression,
-        declaration::{CreateTypeId, EnforceTypeArguments, Monomorphize},
-    },
-    type_engine::*,
-    CallPath, CompileResult, Ident, TypeInfo, TypedDeclaration, TypedFunctionDeclaration,
+    error::*, semantic_analysis::*, type_engine::*, types::*, CallPath, CompileResult, Ident,
+    TypeInfo, TypedDeclaration, TypedFunctionDeclaration,
 };
 
 use super::{module::Module, namespace::Namespace, Path};
