+++ sway-core/src/type_engine/copy_types.rs
@@ -1,0 +1,5 @@
+use super::TypeMapping;
+
+pub(crate) trait CopyTypes {
+    fn copy_types(&mut self, type_mapping: &TypeMapping);
+}
