+++ sway-core/src/type_engine/resolved_type.rs
@@ -13,29 +13,35 @@ pub enum ResolvedType {
     Unit,
     Byte,
     B256,
+    #[allow(dead_code)]
     Struct {
         name: Ident,
         fields: Vec<TypedStructField>,
     },
+    #[allow(dead_code)]
     Enum {
         name: Ident,
         variant_types: Vec<ResolvedType>,
     },
     /// Represents the contract's type as a whole. Used for implementing
     /// traits on the contract itself, to enforce a specific type of ABI.
+    #[allow(dead_code)]
     Contract,
     /// Represents a type which contains methods to issue a contract call.
     /// The specific contract is identified via the `Ident` within.
+    #[allow(dead_code)]
     ContractCaller {
         abi_name: CallPath,
         #[derivative(PartialEq = "ignore", Hash = "ignore")]
         address: Box<TypedExpression>,
     },
+    #[allow(dead_code)]
     Function {
         from: Box<ResolvedType>,
         to: Box<ResolvedType>,
     },
     /// used for recovering from errors in the ast
+    #[allow(dead_code)]
     ErrorRecovery,
 }
 
@@ -56,6 +62,7 @@ impl ResolvedType {
         )
     }
 
+    #[allow(dead_code)]
     pub fn is_numeric(&self) -> bool {
         matches!(self, ResolvedType::UnsignedInteger(_))
     }
