--- sway-core/src/types.rs
@@ -1,3 +1,0 @@
-//! Contains post-type-engine resolved types.
-mod resolved_type;
-pub use resolved_type::ResolvedType;
