+++ sway-core/src/types/friendly_type_string.rs
@@ -1,0 +1,3 @@
+pub(crate) trait FriendlyTypeString {
+    fn friendly_type_str(&self) -> String;
+}
