+++ sway-core/src/types/json_abi_string.rs
@@ -1,0 +1,3 @@
+pub(crate) trait JsonAbiString {
+    fn json_abi_str(&self) -> String;
+}
