+++ sway-core/src/types/mod.rs
@@ -1,0 +1,7 @@
+mod friendly_type_string;
+mod json_abi_string;
+mod to_json_abi;
+
+pub(crate) use friendly_type_string::*;
+pub(crate) use json_abi_string::*;
+pub(crate) use to_json_abi::*;
