+++ sway-core/src/types/to_json_abi.rs
@@ -1,0 +1,5 @@
+use fuels_types::Property;
+
+pub(crate) trait ToJsonAbi {
+    fn generate_json_abi(&self) -> Option<Vec<Property>>;
+}
