+++ test/src/e2e_vm_tests/test_programs/should_fail/unexpected_generic_in_impl/Forc.lock
@@ -1,0 +1,11 @@
+[[package]]
+name = 'core'
+dependencies = []
+
+[[package]]
+name = 'std'
+dependencies = ['core']
+
+[[package]]
+name = 'unexpected_generic_in_impl'
+dependencies = ['std']
