+++ test/src/e2e_vm_tests/test_programs/should_fail/unexpected_generic_in_impl/Forc.toml
@@ -1,0 +1,8 @@
+[project]
+name = "unexpected_generic_in_impl"
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+
+[dependencies]
+std = { path = "../../../../../../sway-lib-std" }
